split up Main component into related pieces
